LegalMindAI/LM_Phase_1/requirements.txt
# Core AI & Orchestration
langgraph>=0.2.0
langchain-google-genai>=1.0.0  # For embeddings only
langchain-pinecone>=0.0.3
pinecone[grpc]>=5.0.0
google-genai>=0.2.0  # 2025 Google GenAI SDK for LLM calls

# Data Ingestion & Utilities
juriscraper>=2.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# LangChain Core
langchain>=0.1.0
langchain-community>=0.0.20

# UI & Memory (Week 4)
streamlit>=1.32.0
psycopg2-binary>=2.9.9

